Implement `network_mode`, `pid` and `ipc` attributes
We use a heavy-duty macro to mass-produce these types, because there are
lot of them in the current format, and the `docker run` command has
another bunch that will inevitably be added at some point.

We're paying the cost of one bit of fairly complex code in exchange for
not needing to hand-roll each of these variants.