v0.3.2: Get rid of auto list-to-string conversion
`docker-compose.yml` contains a number of values which can be serialized
either as a list or as a single string.  See:

https://docs.docker.com/compose/extends/#/adding-and-overriding-configuration

These are theoretically merged as lists when using `extends` and
override files.  But it's weird to serialize them as strings if they're
going to be merged as lists, and it seems like the kind of thing which
might hit a bug somewhere.  So let's just get rid of this particular bit
of cleverness, and always serialize these as lists.