THE `?` OPERATOR HAS LANDED!
See https://blog.rust-lang.org/2016/11/10/Rust-1.13.html

As the release notes indicate, the code:

    try!(try!(try!(foo()).bar()).baz())

...becomes:

    foo()?.bar()?.baz()?

This doesn't actually make a huge difference for `compose_yml`, which is
a bit of an odd library, with relatively little code outside
serde (de)serializers.  But it's going to be fun for some other
libraries.